Extract first measurement after/last measurement before a specific event, during a timeperiod

Extract the last measurement before or the first measurement after a specific other event, during a timepoint.

Usage

first_during_after_event(..., event_datetime, names_suffix = "")

last_during_before_event(..., event_datetime, names_suffix = "")

Arguments

...

Arguments passed on to summarise_during

x

An adm data frame

y

A (tidy) data frame, with a datetime column

datetime

The column of y to use as the main datetime for matching

during

The time period to extract data for, one of:

"during_visit"

"during_visit_initial_24h"

"during_visit_initial_72h

"during_icu"

"before_visit_initial_24h"

"14_days_before_visit_until_visit_end"

"year_before_visit_until_visit_end"

"before_visit_end"

"30_days_before_visit_start"

"year_before_initial_24h"

"during_value"

event_datetime

The datetime that the extraction should be relative to.

names_suffix

A text label to use as the end of the column name created.

Examples

#' # raw data that we will combine
adm_data_example
#> # A tibble: 5 × 13
#>   person_id visit_id visit_start_datetime visit_end_datetime  gender ethnicity death_date age_at_visit_start adm_service        ward    discharge_destination discharged_deceased readmitted_within_30…¹
#>   <chr>        <int> <dttm>               <dttm>              <chr>  <chr>     <dttm>                  <int> <chr>              <chr>   <chr>                 <chr>               <chr>                 
#> 1 AA               1 2021-01-01 09:00:00  2021-01-03 18:00:00 Female "White\r… NA                         45 Emergency Medicine EMERGE… 0                     0                   1                     
#> 2 AA               2 2021-01-05 11:00:00  2021-01-07 18:00:00 Female "White\r… NA                         45 Emergency Medicine EMERGE… 0                     1                   0                     
#> 3 BB               3 2021-01-20 09:00:00  2021-02-01 18:00:00 Male   "White\r… NA                         12 Nephrology         DIALYS… 1                     0                   0                     
#> 4 CC               4 2021-01-01 08:00:00  2021-02-01 18:00:00 Female "White\r… NA                         50 Emergency Medicine EMERGE… 0                     0                   1                     
#> 5 CC               5 2021-02-02 08:00:00  2021-02-02 18:00:00 Female "White\r… NA                         50 Emergency Medicine EMERGE… 0                     0                   0                     
#> # ℹ abbreviated name: ¹​readmitted_within_30_days
fsheet_news2_example
#> # A tibble: 12 × 17
#>    person_id symbol measurement_datetime value_as_number censoring value_as_logical value_as_character name        unit  type    template  form         data_id line_id measurement_id title comment
#>    <chr>     <chr>  <dttm>                         <dbl> <chr>     <lgl>            <chr>              <chr>       <chr> <chr>   <chr>     <chr>          <int>   <int> <chr>          <chr> <chr>  
#>  1 CC        news2  2021-01-01 08:59:00                5 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  2 AA        news2  2021-01-01 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  3 AA        news2  2021-01-01 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  4 AA        news2  2021-01-01 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  5 CC        news2  2021-01-03 17:00:00                7 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  6 CC        news2  2021-01-04 09:00:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  7 AA        news2  2021-01-05 11:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  8 AA        news2  2021-01-05 16:30:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  9 AA        news2  2021-01-06 10:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 10 BB        news2  2021-01-21 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 11 BB        news2  2021-01-22 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 12 BB        news2  2021-01-23 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     

# extract the first NEWS2 record during each hospital visit
first_news2 <- adm_data_example %>%
  first_during(fsheet_news2_example,
             datetime = measurement_datetime,
             during = "during_visit") %>%
  select(person_id,
         visit_id,
         visit_start_datetime,
         visit_end_datetime,
         contains("news2"))
#> ℹ Extracting all during
#> ℹ Checking datetime is not all midnight
#> ✔ Checking datetime is not all midnight [15ms]
#> 
#> ℹ Extracting all during
#> ℹ Joining tables
#> ✔ Joining tables [27ms]
#> 
#> ℹ Extracting all during
#> ✔ Extracting all during [70ms]
#> 
#> ℹ Summarising/slicing
#> ✔ Summarising/slicing [20ms]
#> 
#> ℹ Pivoting wider
#> ℹ Identifying symbol-type pairs
#> ✔ Identifying symbol-type pairs [16ms]
#> 
#> ℹ Pivoting wider
#> ℹ Pivoting wider
#> ✔ Pivoting wider [34ms]
#> 
#> ℹ Pivoting wider
#> ℹ Relocating columns
#> ✔ Relocating columns [215ms]
#> 
#> ℹ Pivoting wider
#> ✔ Pivoting wider [286ms]
#> 

# extract the first SpO2 measurement that FOLLOW the first NEWS2 record
# during each hospital visit
first_news2 %>%
 first_during_after_event(fsheet_spo2_example,
                          datetime = measurement_datetime,
                          during = "during_visit",
                          event_datetime = news2_first_during_visit_datetime,
                          names_suffix = "first_news2") %>%
  select(person_id,
         visit_id,
         contains("spo2"),
         contains("news2"))
#> ℹ Extracting all during
#> ℹ Checking datetime is not all midnight
#> ✔ Checking datetime is not all midnight [8ms]
#> 
#> ℹ Extracting all during
#> ℹ Joining tables
#> ✔ Joining tables [20ms]
#> 
#> ℹ Extracting all during
#> ✔ Extracting all during [42ms]
#> 
#> ℹ Summarising/slicing
#> ✔ Summarising/slicing [26ms]
#> 
#> ℹ Pivoting wider
#> ℹ Identifying symbol-type pairs
#> ✔ Identifying symbol-type pairs [14ms]
#> 
#> ℹ Pivoting wider
#> ℹ Pivoting wider
#> ✔ Pivoting wider [33ms]
#> 
#> ℹ Pivoting wider
#> ℹ Relocating columns
#> ✔ Relocating columns [17ms]
#> 
#> ℹ Pivoting wider
#> ✔ Pivoting wider [88ms]
#> 
#> # A tibble: 5 × 6
#>   person_id visit_id spo2_first_after_first_news2_during_visit_datetime spo2_first_after_first_news2_during_visit news2_first_during_visit_datetime news2_first_during_visit
#>   <chr>        <int> <dttm>                                                                                 <dbl> <dttm>                                               <dbl>
#> 1 AA               1 2021-01-01 10:59:00                                                                       96 2021-01-01 09:00:00                                      3
#> 2 AA               2 2021-01-05 11:30:00                                                                       96 2021-01-05 11:30:00                                      3
#> 3 BB               3 2021-01-23 08:59:00                                                                       85 2021-01-21 09:30:00                                      3
#> 4 CC               4 NA                                                                                        NA 2021-01-01 08:59:00                                      5
#> 5 CC               5 NA                                                                                        NA NA                                                      NA

# extract the LAST SpO2 measurement that is before (or the same time as)
# the first NEWS2 record during each hospital visit
first_news2 %>%
 last_during_before_event(fsheet_spo2_example,
                          datetime = measurement_datetime,
                          during = "during_visit",
                          event_datetime = news2_first_during_visit_datetime,
                          names_suffix = "first_news2") %>%
  select(person_id,
         visit_id,
         contains("spo2"),
         contains("news2"))
#> ℹ Extracting all during
#> ℹ Checking datetime is not all midnight
#> ✔ Checking datetime is not all midnight [8ms]
#> 
#> ℹ Extracting all during
#> ℹ Joining tables
#> ✔ Joining tables [20ms]
#> 
#> ℹ Extracting all during
#> ✔ Extracting all during [41ms]
#> 
#> ℹ Summarising/slicing
#> ✔ Summarising/slicing [24ms]
#> 
#> ℹ Pivoting wider
#> ℹ Identifying symbol-type pairs
#> ✔ Identifying symbol-type pairs [13ms]
#> 
#> ℹ Pivoting wider
#> ℹ Pivoting wider
#> ✔ Pivoting wider [33ms]
#> 
#> ℹ Pivoting wider
#> ℹ Relocating columns
#> ✔ Relocating columns [17ms]
#> 
#> ℹ Pivoting wider
#> ✔ Pivoting wider [87ms]
#> 
#> # A tibble: 5 × 6
#>   person_id visit_id spo2_last_before_first_news2_during_visit_datetime spo2_last_before_first_news2_during_visit news2_first_during_visit_datetime news2_first_during_visit
#>   <chr>        <int> <dttm>                                                                                 <dbl> <dttm>                                               <dbl>
#> 1 AA               1 NA                                                                                        NA 2021-01-01 09:00:00                                      3
#> 2 AA               2 2021-01-05 11:30:00                                                                       96 2021-01-05 11:30:00                                      3
#> 3 BB               3 2021-01-21 09:15:00                                                                       93 2021-01-21 09:30:00                                      3
#> 4 CC               4 NA                                                                                        NA 2021-01-01 08:59:00                                      5
#> 5 CC               5 NA                                                                                        NA NA                                                      NA