Basic cleanup, checks, and annotation of ADT data

Performs basic cleanup, checks and annotation of ADT data with indexes for the hospital visit (visit_index) and the visit to each departments (department_visit_index), and basic labels for each department.

Usage

adt_annotate(x, fixed_labels, annotate_fn)

Arguments

x

Tidy ADT data, as tidied by adt_rename()

fixed_labels

A data frame with a colum department, which will be joined to the ADT data

annotate_fn

A function that provides additional annotation

Value

The supplied tidy ADT data x with additional annotations

Details

The clean and checks performed are:

Removes duplicate rows

Checks discharge dates are identical within a visit

The annotations added are:

Labels hospital visits via visit_index

Adds the end_datetime for each row of ADT (which is not always included in the raw data)

Labels visits to specific departments. A visit to a single department may involve multiple rows of ADT data, since a patient may move beds within a department one or more times. The visits are labelled with department_visit_index, and the corresponding times recorded in department_start_datetime and department_end_datetime, and department_length_days.

(Left) joins in fixed_labels

Runs the user-supplied annotation function annotate_fn

See also

Author

R.J.B. Goudie

Examples

fixed_labels <-
  tibble::tribble(
    ~department, ~is_ed,
    "ADD EMERGENCY DEPT", TRUE,
    "ADD C2 WARD", FALSE,
    "ADD D2 WARD", FALSE,
    "DIALYSIS CENTRE", FALSE,
    "ADD A1 WARD", FALSE,
    "ADD A2 WARD", FALSE,
    "ADD A3 WARD", FALSE,
    "ADD A4 WARD", FALSE,
    "ADD X WARD", FALSE,
    "ADD Y WARD", FALSE,
    "ADD Z WARD", FALSE)

adt_data_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("adt.csv"),
    col_types = "adt"
  ) %>%
  adt_rename
#> Warning: The following named parsers don't match the column names: ADT_SERV_AREA_NAME, HOSP_SERV_NAME
#> Warning: One or more parsing issues, call `problems()` on your data frame for details, e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> ! Some columns were missing: ADT_SERV_AREA_NAME and HOSP_SERV_NAME

adt_data_example %>%
  adt_annotate(annotate_fn = identity,
               fixed_labels = fixed_labels)
#> # A tibble: 22 × 16
#>    person_id visit_id event_type  visit_index start_datetime      discharge_datetime  end_datetime        department     bed   room  department_just_moved department_visit_index department_start_dat…¹
#>    <chr>        <int> <chr>             <int> <dttm>              <dttm>              <dttm>              <chr>          <chr> <chr> <lgl>                                  <int> <dttm>                
#>  1 CC               4 Admission             1 2021-01-01 08:00:00 2021-02-01 18:00:00 2021-01-01 09:00:00 ADD EMERGENCY… A-01  NA    TRUE                                       1 2021-01-01 08:00:00   
#>  2 AA               1 Admission             1 2021-01-01 09:00:00 2021-01-03 18:00:00 2021-01-01 12:00:00 ADD EMERGENCY… A-01  NA    TRUE                                       1 2021-01-01 09:00:00   
#>  3 CC               4 Transfer In           1 2021-01-01 09:00:00 2021-02-01 18:00:00 2021-01-05 09:00:00 ADD X WARD     X-01  NA    TRUE                                       2 2021-01-01 09:00:00   
#>  4 AA               1 Transfer In           1 2021-01-01 12:00:00 2021-01-03 18:00:00 2021-01-02 09:00:00 ADD C2 WARD    B-01  NA    TRUE                                       2 2021-01-01 12:00:00   
#>  5 AA               1 Transfer In           1 2021-01-02 09:00:00 2021-01-03 18:00:00 2021-01-03 18:00:00 ADD C2 WARD    B-02  NA    FALSE                                      2 2021-01-01 12:00:00   
#>  6 AA               1 Discharge             1 2021-01-03 18:00:00 2021-01-03 18:00:00 NA                  NA             NA    NA    TRUE                                       3 2021-01-03 18:00:00   
#>  7 CC               4 Transfer In           1 2021-01-05 09:00:00 2021-02-01 18:00:00 2021-01-10 09:00:00 ADD X WARD     X-02  NA    FALSE                                      2 2021-01-01 09:00:00   
#>  8 AA               2 Admission             2 2021-01-05 11:00:00 2021-01-07 18:00:00 2021-01-05 12:00:00 ADD EMERGENCY… A-01  NA    TRUE                                       1 2021-01-05 11:00:00   
#>  9 AA               2 Transfer In           2 2021-01-05 12:00:00 2021-01-07 18:00:00 2021-01-07 18:00:00 ADD D2 WARD    D-01  NA    TRUE                                       2 2021-01-05 12:00:00   
#> 10 AA               2 Discharge             2 2021-01-07 18:00:00 2021-01-07 18:00:00 NA                  NA             NA    NA    TRUE                                       3 2021-01-07 18:00:00   
#> # ℹ 12 more rows
#> # ℹ abbreviated name: ¹​department_start_datetime
#> # ℹ 3 more variables: department_end_datetime <dttm>, department_length_days <dbl>, is_ed <lgl>