Add a new definition of a Diagnosis PL item

Add a new definition of a Diagnosis PL item

Extract condition data into tidy format

Usage

condition_add(
  condition_def,
  symbol,
  title,
  icd10 = NA,
  condition_start_datetime_fn = case_when(!is.na(diagnosis_datetime) ~
    diagnosis_datetime, !is.na(entered_datetime) ~ entered_datetime, TRUE ~ ymd_hms(NA)),
  condition_end_datetime_fn = case_when(!is.na(resolved_datetime) ~ resolved_datetime,
    TRUE ~ ymd_hms(NA)),
  silently_exclude_deleted_when = TRUE,
  silently_exclude_resolved_when = FALSE
)

condition_extract(x, condition_def, errors = stop)

Arguments

condition_def

A condition definition

symbol

A short name for the condition item. This should be a character vector that is suitable to use as an R object name

title

The full name for the condition item, suitable for using in figure titles etc

x

condition data in tidy-column-name format (after applying condition_rename)

errors

A function indicating what to do when an error is found

icd_10

A vector of ICD-10 codes to use.

silently_exclude_deleted

When TRUE, diagnoses with status = "Deleted" will be removed from the data frame. In theory, only diagnoses made in error should be deleted, but in practice it is thought that clinical staff often delete diagnoses that are resolved rather than mark them as "Resolved".

Value

A data frame with the following columns: person_id, symbol, value_as_number, value_as_character, censoring, ordered_datetime, collected_datetime, received_datetime, result_datetime, ordering_department_name, range_low, range_high, unit, name, title, method, group, type

The result will be sorted by diagnosis_datetime

Author

R.J.B. Goudie