Load data from Epic in delimited format (including CSV) into a database

Reads in data in a chunked manner, and writes it to a database using DBI::dbWriteTable

Usage

db_write_tidybrookes_csv(
  file,
  connection,
  table_name,
  col_types,
  na = c("", "NA"),
  quote = "\"",
  progress = TRUE
)

db_write_tidybrookes_delim(
  file,
  connection,
  table_name,
  col_types,
  na = c("", "NA"),
  quote = "\"",
  ...
)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector), as per readr::read_delim()

connection

A DBIConnection object, as returned by DBI::dbConnect()

table_name

A table name, passed to DBI::dbWriteTable()

col_types

Either a string specifying the name of the table that the data comes from, for which a standard format can be used. Available values are adm, adt, demogs, fsheet, fsheet_io, tests, med_admin, med_prescr, diagnosis_pl, med_hist, radiology. This is translated into column specification using default_col_types.

Or a readr::cols() format list of column names and types, which can be used where a nonstandard data format are supplied.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

quote

Single character used to quote strings