Load data from CUH in delimited (including CSV) format into duckdb

Convenience wrapper that makes it easier to set the col_types for standard data from CUH for loading into duckdb.

Usage

duckdb_read_tidybrookes_csv(
  src,
  table,
  file,
  col_types = table,
  na = c("", "NA"),
  quote = "\"",
  table_name = paste0(table, "_raw_unrenamed"),
  strict_mode = TRUE,
  append = FALSE
)

Arguments

src

A DBIConnection object produced by duckdb_tidybrookes_connect or DBI::dbConnect()

table

A string specifying the name of the table that the data comes from, for which a standard format can be used. Available values are adm, adt, demogs, fsheet, fsheet_io, tests, med_admin, med_prescr, diagnosis_pl, med_hist, radiology. This is translated into column specification using default_col_types.

file

A path to the CSV file

col_types

Either a string specifying the name of the table that the data comes from, for which a standard format can be used. Available values are adm, adt, demogs, fsheet, fsheet_io, tests, med_admin, med_prescr, diagnosis_pl, med_hist, radiology. This is translated into column specification using default_col_types.

Or a readr::cols() format list of column names and types, which can be used where a nonstandard data format are supplied.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

quote

Single character used to quote strings

table_name

The name to use for the table in duckdb

strict_mode

Should duckdb's strict mode for loading CSVs be used?

append

Should new data be appended to the end of an existing table in the database?

Author

R.J.B. Goudie