Connect to duckdb database with suitable defaults

This is a wrapper around DBI::dbConnect() that sets suitable defaults for use with tidybrookes datasets:

uses bit64 integer64 representation for big integers, since some IDs from EHRs can be larger than can be represented by standard integers and representing as floats/doubles is incorrect

Since duckdb support for timezone is not great in R currently, we store datetimes in UTC in duckdb and force conversion to local time at export from duckdb.

Usage

duckdb_tidybrookes_connect(
  db_dir = "",
  bigint = "integer64",
  timezone_out = "Europe/London",
  tz_out_convert = "force",
  temp_dir = NULL
)

Arguments

db_dir

Location for database files. Should be a path to an existing directory in the file system. With the default (or ""), all data is kept in RAM.

timezone_out

The time zone returned to R, defaults to "UTC", which is currently the only timezone supported by duckdb. If you want to display datetime values in the local timezone, set to Sys.timezone() or "".

tz_out_convert

How to convert timestamp columns to the timezone specified in timezone_out. There are two options: "with", and "force". If "with" is chosen, the timestamp will be returned as it would appear in the specified time zone. If "force" is chosen, the timestamp will have the same clock time as the timestamp in the database, but with the new time

temp_dir

Set the directory to which duckdb should write temp files. For data stored on slow network drives, it may be much faster to set this to a faster local drive.

Details

Facility to specify the temp_directory for duckdb is also provided

Author

R.J.B. Goudie

Examples

if (FALSE) { # \dontrun{
con <- duckdb_tidybrookes_connect()
} # }