Add and extract flowsheet items

fsheet_add creates a new "fsheet definition", which is a specification of how the raw flowsheet data should be extracted, filtered, cleaned and standardised. A list of such specifications can be used on raw fsheet data using fsheet_extract.

Usage

fsheet_add(
  fsheet_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  type = "numeric",
  silently_exclude_na_when = FALSE,
  silently_exclude_when = FALSE,
  censoring_fn = case_when(TRUE ~ NA_character_),
  value_as_number_fn = if_else(type == "numeric", case_when(TRUE ~ value_as_number),
    case_when(TRUE ~ NA_real_)),
  value_as_character_fn = if_else(type == "character", case_when(TRUE ~
    value_as_character), case_when(TRUE ~ NA_character_)),
  value_as_logical_fn = if_else(type == "logical", case_when(TRUE ~ value_as_logical),
    case_when(TRUE ~ as.logical(NA))),
  unit_rescale_fn = case_when(TRUE ~ value_as_number),
  unit_relabel_fn = case_when(TRUE ~ NA_character_),
  coalesce_fn = identity,
  expect_before = TRUE,
  expect_after = TRUE,
  range_mainly_low = NA,
  range_mainly_high = NA,
  range_discard_below = NA,
  range_discard_above = NA
)

fsheet_extract(
  x,
  fsheet_def,
  errors = stop,
  rds_only = FALSE,
  rds_filepath_fn = NULL
)

Arguments

fsheet_def

A list of fsheet definitions

symbol

A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name

title

The full name for the flowsheet item, suitable for using in figure titles etc

names

A character vector of display names (disp_name in raw format) of the flowsheet items to include

search_pattern

A character vector of potential synonymns for this flowsheet item, in case new names (disp_name in raw format) for the same item are added in future. These will be searched for in the raw data and if any potential new names are identified, the user will be alerted to the presence of potential new flowsheet names.

search_exclude

A character vector of flowsheet item names (disp_name in raw format) that are known NOT to be relevant to this flowsheet item.

type

Whether flowsheet item is a number or a string. Either "numeric" or "character"

silently_exclude_na_when

An expression that returns a logical value, specifying the circumstances when NA values should be automatically excluded.

silently_exclude_when

An expression that returns a logical value, specifying the circumstances when values should be automatically excluded

censoring_fn

A function specifying the value of the censoring column

value_as_number_fn

A function specifying how to generate the value_as_number column

value_as_character_fn

A function specifying how to generate the value_as_character column

value_as_logical_fn

A function specifying how to generate the value_as_logical column

unit_rescale_fn

A function specifying how to rescale the flowsheet values. By default, does not rescale

unit_relabel_fn

A function specifying how to relabel the units of the the flowsheet values. By default, does not rescale

coalesce_fn

A function specifying how to combine measurements from exactly the same time

expect_before

An expression that returns a logical value, specifying a condition that should be TRUE in the raw data

expect_after

An expression that returns a logical value, specifying a condition that should be TRUE in the processed (output) data

range_mainly_low, range_mainly_high

An indicative lower and upper range for most values. Values outside this range are NOT excluded: this is purely for setting default scales of plots etc

range_discard_below, range_discard_above

This pair set the lower and upper range for the flowsheet item. Values outside this range are EXCLUDED.

x

Flowsheet data in renamed format (after applying fsheet_rename)

errors

A function indicating what to do when an error is found

Value

fsheet_add returns a list of fsheet definitions, with the new definition added to the end of the supplied fsheet_def

fsheet_extract returns a data frame with the following columns: person_id, symbol, value_as_character, value_as_number, value_as_logical, censoring, comment, measurement_datetime, name, title, data_id, measurement_id, line_id, template, form, type, unit

Details

Flowsheet defintiions describe how the raw data should be extracted, filtered, cleaned and standardised as follows:

Labelling the flowsheet item. Two names/labels are required each fsheet item:

a symbol, which serves as the "R variable name" for the item. To make the symbol quick and easy to type, and work with in R, it is recommended that these are lowercase, ideally single words (e.g. "news2" or "spo2)

a title, which serves as the label the of the flowsheet item in plots and figures. This should be capitalised as you would want the item to be labelled in formal documents (e.g. "NEWS2" or "SpO2).

Extracting relevant flowsheet items The specification of which rows should be extracted is via:

names, a vector listing the labels of the fsheet items in the raw data. This corresponds to column labelled as name by fsheet_rename, which is labelled as disp_name in the raw extracts from Epic).

Additionally, to guard against the potential for new names for flowsheet items being introduced through the evolution of the EHR data, search terms for potential other names for this flowsheet items (and known exclusions) can be provided using:

search_pattern, a vector of search terms of potential synonyms for the flowsheet item. This list of search terms will be searched whenever the flowsheet item is extracted, and if a possible new data field is identified, then this will be alerted to the user.

search_exclude, a vector of names of flowsheet items that are known not to be relevant. This list of names will be removed prior to searching the data for the search patterns in search_pattern. This helps to avoid alerting the user about numerous flowsheet items that are known to be irrelevant.

Excluding rows. Rows containing NA in the value column may be always irrelevant, so it may be helpful to remove these by setting silently_exclude_na_when to TRUE.

Alternatively, an expression that describes when row should be excluded can be provided using silently_exclude_when. For example, setting this to (value_original == "Other (Comment)") will exclude all rows where the value is "Other (Comment)".

For numerical flowsheet items,range_discard_below and range_discard_above can be specified so that values outside this range are excluded (see below for details).

Converting values to numbers/characters/logicals. Raw flowsheet data are a complex mix of numerical data (sometimes containing censoring, and other non-numerical data), categorical data (coded as characters) and logical data (usually coded as characters). Standardising the raw data into data of a single type makes downstream analysis easier, so several aspects of the definition specify how this is done:

type: one of numeric, character or logical. This is used to set sensible defaults for other aspects of the specification, and is added as a column to the output, so downstream handling of the data can use this column to determine handling of the data (e.g. default plots might be a barplot for character items and a lineplot for numeric items)

value_as_number_fn, value_as_character_fn and value_as_logical_fn enable functions that handle conversion of the raw data. TODO add more details

unit_rescale_fn and unit_relabel_fn enable functions that standardise the units to be provided

coalesce_fn enables a function that combines together data to be provided. For instance, this can be used to specify how simultaneous data should be combined: e.g. how should two blood pressure records at the exact same time be handled? The function can describe how to combine the information by e.g. taking the mean or max, or arbitrarily choosing one of the records.

Validating that data conform to expectations. Two parameters enable checking that the data comply with known restrictions. For example, we may know a finite range exists for a flowsheet item, and we want to check that all data fall within this range.

Alternatively, we may need to map each value of the raw data to a smaller range of values, and we need to be sure that our function for doing this handles all the values that are present in the raw data (i.e. no new values have been added).

Checks that should be performed on the raw data should be supplied as expect_before, whereas checks that should be performed on the cleaned data should be supplied as expect_after.

Specification of ranges. Two types of ranges can be specified:

range_discard_below and range_discard_above: these are intended to describe hard limits: values outside this range will be excluded.

range_mainly_low and range_mainly_high: these are intended to describe the range of values we typically expect most of the data to fall within, which can be useful for setting default ranges for e.g. default plots. Values outside this range are not excluded.

Author

R.J.B. Goudie

Examples

# A definition for the NEWS2 score:
# - Includes the value of the flowsheet called "NEWS 2 score"
# - Excludes all NA values, and also " " and "E".
# - The value_as_number_fn will convert "3mm" to 3, since this is probably
#   a typo
# - Checks that all output values are in integers between 0 and 17.
fsheet_def <- list() %>%
  fsheet_add(
    symbol = "news2",
    title = "NEWS2",
    names = c("NEWS2 score"),
    search_pattern = c("news2", "news"),
    search_exclude = c(),
    silently_exclude_na_when = TRUE,
    silently_exclude_when =
      (value_original == " " |
         value_original == "E" # some kind of typo
      ),
    value_as_number_fn =
      case_when(value_original == "3mm" ~ 3,
                TRUE ~ value_as_number),
    expect_after =
      (value_as_number %in% 0:17))

# Note that person_id == "BB" has a NEWS2 with value="3mm", and that
# person_id == "CC" has a NEWS2 with NA value
fsheet_raw_example %>% print(n = Inf)
#> # A tibble: 31 × 10
#>    person_id  data_id line_id measurement_id measurement_datetime name               value comment template  form       
#>    <chr>        <int>   <int> <chr>          <dttm>               <chr>              <chr> <chr>   <chr>     <chr>      
#>  1 AA               5      61 1234           2021-01-03 17:00:00  Weight             4125  NA      IP Vitals Vital signs
#>  2 AA        10909029      60 10             2021-01-01 10:59:00  SpO2               96    NA      IP Vitals Vital signs
#>  3 AA        10909029      60 10             2021-01-02 08:59:00  SpO2               90    NA      IP Vitals Vital signs
#>  4 AA        10926333     239 1234           2021-01-01 09:30:00  NEWS2 score        3     NA      IP Vitals Vital signs
#>  5 AA        10926333     239 1234           2021-01-01 12:30:00  NEWS2 score        4     NA      IP Vitals Vital signs
#>  6 AA        10926333     239 1234           2021-01-01 09:00:00  NEWS2 score        3     NA      IP Vitals Vital signs
#>  7 AA        10909029      60 10             2021-01-05 11:30:00  SpO2               96    NA      IP Vitals Vital signs
#>  8 AA        10909029      60 10             2021-01-05 16:30:00  SpO2               90    NA      IP Vitals Vital signs
#>  9 AA        10909029      60 10             2021-01-06 10:30:00  SpO2               92    NA      IP Vitals Vital signs
#> 10 AA        10926333     239 1234           2021-01-05 11:30:00  NEWS2 score        3     NA      IP Vitals Vital signs
#> 11 AA        10926333     239 1234           2021-01-05 16:30:00  NEWS2 score        6     NA      IP Vitals Vital signs
#> 12 AA        10926333     239 1234           2021-01-06 10:30:00  NEWS2 score        3     NA      IP Vitals Vital signs
#> 13 BB        10926333     239 1234           2021-01-19 08:59:00  Weight             4000  NA      IP Vitals Vital signs
#> 14 BB               5      61 1234           2021-01-20 09:00:00  Weight             4035  NA      IP Vitals Vital signs
#> 15 BB               5      61 1234           2021-01-20 08:59:00  Weight             4025  NA      IP Vitals Vital signs
#> 16 BB               5      61 1234           2021-01-25 08:59:00  Weight             4000  NA      IP Vitals Vital signs
#> 17 BB        10909029      60 10             2021-01-24 09:00:00  NEWS2/MEOWS: ACVPU A     NA      IP Vitals Vital signs
#> 18 BB        10909029      60 10             2021-01-24 09:00:00  NEWS2/MEOWS: ACVPU P     NA      IP Vitals Vital signs
#> 19 BB        10926333     239 1234           2021-01-21 09:30:00  NEWS2 score        3     NA      IP Vitals Vital signs
#> 20 BB        10926333     239 1234           2021-01-22 12:30:00  NEWS2 score        4     NA      IP Vitals Vital signs
#> 21 BB        10926333     239 1234           2021-01-23 09:00:00  NEWS2 score        3mm   NA      IP Vitals Vital signs
#> 22 BB        10909029      60 10             2021-01-21 08:59:00  SpO2               93    NA      IP Vitals Vital signs
#> 23 BB        10909029      60 10             2021-01-23 08:59:00  SpO2               85    NA      IP Vitals Vital signs
#> 24 BB        10909029      60 10             2021-01-26 12:59:00  SpO2               80    NA      IP Vitals Vital signs
#> 25 BB        10909029      60 10             2021-02-01 19:00:00  SpO2               30    NA      IP Vitals Vital signs
#> 26 CC        10926333     239 1234           2021-01-01 08:59:00  NEWS2 score        5     NA      IP Vitals Vital signs
#> 27 CC        10926333     239 1234           2021-01-03 17:00:00  NEWS2 score        7     NA      IP Vitals Vital signs
#> 28 CC        10926333     239 1234           2021-01-04 09:00:00  NEWS2 score        6     NA      IP Vitals Vital signs
#> 29 CC        10926333     239 1234           2021-01-02 08:59:00  NEWS2 score        NA    NA      IP Vitals Vital signs
#> 30 CC        10909029      60 10             NA                   SpO2               93    NA      IP Vitals Vital signs
#> 31 NA              NA      NA NA             NA                   NA                 NA    NA      NA        NA         

# Note that the NA row is excluded, and that "3mm" is converted to 3
fsheet_extract(fsheet_raw_example, fsheet_def)
#> ℹ Extracting NEWS2 (news2) from 13 raw rows
#> ℹ 1 row removed since value was NA
#> ℹ 12 rows extracted
#> # A tibble: 12 × 17
#>    person_id symbol measurement_datetime value_as_number censoring value_as_logical value_as_character name        unit  type    template  form         data_id line_id measurement_id title comment
#>    <chr>     <chr>  <dttm>                         <dbl> <chr>     <lgl>            <chr>              <chr>       <chr> <chr>   <chr>     <chr>          <int>   <int> <chr>          <chr> <chr>  
#>  1 CC        news2  2021-01-01 08:59:00                5 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  2 AA        news2  2021-01-01 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  3 AA        news2  2021-01-01 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  4 AA        news2  2021-01-01 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  5 CC        news2  2021-01-03 17:00:00                7 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  6 CC        news2  2021-01-04 09:00:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  7 AA        news2  2021-01-05 11:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  8 AA        news2  2021-01-05 16:30:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  9 AA        news2  2021-01-06 10:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 10 BB        news2  2021-01-21 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 11 BB        news2  2021-01-22 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 12 BB        news2  2021-01-23 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA