Add a new definition of a flowsheet item

Add a new definition of a flowsheet item

Extract fsheet_io data into tidy format

Usage

fsheet_io_add(
  fsheet_io_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  type = "numeric",
  silently_exclude_na_when = FALSE,
  silently_exclude_when = FALSE,
  censoring_fn = case_when(TRUE ~ NA_character_),
  value_as_number_fn = if_else(type == "numeric", case_when(TRUE ~ value_as_number),
    case_when(TRUE ~ NA_real_)),
  value_as_character_fn = if_else(type == "character", case_when(TRUE ~
    value_as_character), case_when(TRUE ~ NA_character_)),
  value_as_logical_fn = if_else(type == "logical", case_when(TRUE ~ value_as_logical),
    case_when(TRUE ~ as.logical(NA))),
  unit_rescale_fn = case_when(TRUE ~ value_as_number),
  unit_relabel_fn = case_when(TRUE ~ NA_character_),
  coalesce_fn = identity,
  expect_before = TRUE,
  expect_after = TRUE,
  range_mainly_low = NA,
  range_mainly_high = NA,
  range_discard_below = NA,
  range_discard_above = NA
)

fsheet_io_extract(x, fsheet_io_def, errors = stop)

Arguments

fsheet_io_def

A fsheet_io definition

symbol

A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name

title

The full name for the flowsheet item, suitable for using in figure titles etc

names

The display names (disp_name in raw format) of the flowsheet items to include

search_pattern

A charater vector of potential synonymns for this flowsheet item, in case new names (disp_name in raw format) for the same item are added in future.

search_exclude

A character vector of flowsheet item names (disp_name in raw format) that are known NOT to be relevant to this flowsheet item.

type

Whether flowsheet item is a number or a string. Either "numeric" or "character"

silently_exclude_na_when

An expression that returns a logical value, specifying the circumstances when NA values should be automatically excluded.

silently_exclude_when

An expression that returns a logical value, specifying the circumstances when values should be automatically excluded

censoring_fn

A function specifying the value of the censoring column

value_as_number_fn

A function specifying how to generate the value_as_number column

value_as_character_fn

A function specifying how to generate the value_as_character column

value_as_logical_fn

A function specifying how to generate the value_as_logical column

unit_rescale_fn

A function specifying how to rescale the flowsheet values. By default, does not rescale

unit_relabel_fn

A function specifying how to relabel the units of the the flowsheet values. By default, does not rescale

coalesce_fn

A function specifying how to combine measurements from exactly the same time

expect_before

An expression that returns a logical value, specifying a condition that should be TRUE in the raw data

expect_after

An expression that returns a logical value, specifying a condition that should be TRUE in the processed (output) data

range_mainly_low, range_mainly_high

An indicative lower and upper range for most values. Values outside this range are NOT excluded: this is purely for setting default scales of plots etc

range_discard_below, range_discard_above

This pair set the lower and upper range for the flowsheet item. Values outside this range are EXCLUDED.

x

Flowsheet data in renamed format (after applying fsheet_io_rename)

errors

A function indicating what to do when an error is found

Value

A data frame with the following columns: person_id, symbol, value_as_character, value_as_number, value_as_logical, censoring, comment, measurement_datetime, name, title, data_id, measurement_id, line_id, template, form, type, unit

Author

R.J.B. Goudie