Add a new definition of a test item

Add a new definition of a test item

Extract med_prescr data into tidy format

Usage

med_prescr_add(
  med_prescr_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  silently_exclude_na_routes = FALSE,
  route,
  route_exclude = NA,
  dose_as_number_fn = case_when(TRUE ~ dose_as_number),
  dose_rescale_fn = case_when(TRUE ~ as.numeric(dose_as_number)),
  unit_relabel_fn = case_when(TRUE ~ unit),
  expect_after = TRUE
)

med_prescr_extract(x, med_prescr_def, errors = stop)

Arguments

med_prescr_def

A med_prescr definition

symbol

A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name

title

The full name for the test item, suitable for using in figure titles etc

names

The names of the Medications (DrugName in raw format)

search_pattern

A charater vector of potential synonymns for this test, in case new names (TestName in raw format) for the same test are added in future.

search_exclude

A character vector of test names (TestName in raw format) that are known NOT to be relevant to this test.

unit_relabel_fn

A function specifying how to relabel the units of the the result values. By default, does not rescale.

expect_after

An expression that returns a logical value, specifying a condition that should be TRUE in the processed (output) data

x

Medication administration data in tidy-column-name format (after applying med_prescr_rename)

errors

A function indicating what to do when an error is found

search_exclude_group

A character vector of test group names (TestGroupName in raw format) that are known NOT to be relevant to this test.

type

Whether flowsheet item is a number or a string. Either "numeric" or "character"

silently_exclude_na_when

An expression that returns a logical value, specifying the circumstances when NA values should be automatically excluded.

silently_exclude_when

An expression that returns a logical value, specifying the circumstances when values should be automatically excluded

censoring_fn

A function specifying the value of the censoring column

value_as_number_fn

A function specifying how to generate the value_as_number column

value_as_character_fn

A function specifying how to generate the value_as_character column

unit_rescale_fn

A function specifying how to rescale the result values. By default, does not rescale.

expect_before

An expression that returns a logical value, specifying a condition that should be TRUE in the raw data

range_mainly_low, range_mainly_high

An indicative lower and upper range for most values. Values outside this range are NOT excluded: this is purely for setting default scales of plots etc

range_discard_below, range_discard_above

This pair set the lower and upper range for the flowsheet item. Values outside this range are EXCLUDED.

Value

A data frame with the following columns: person_id, symbol, value_as_number, value_as_character, censoring, administered_datetime,

unit, name, title, method, group, type

The result will be sorted by administered_datetime

Author

R.J.B. Goudie