Add a new definition of a radiology item

Add a new definition of a radiology item

Extract radiology data into tidy format

Usage

radiology_add(
  radiology_def,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  value_as_character_fn = narrative
)

radiology_extract(x, radiology_def, errors = stop)

Arguments

radiology_def

A radiology definition

symbol

A short name for the radiology item. This should be a character vector that is suitable to use as an R object name

title

The full name for the radiology item, suitable for using in figure titles etc

names

The names of the procedure names (Proc_Name in raw format)

search_pattern

A charater vector of potential synonymns for this radiology item, in case new radiology names (Proc_Name in raw format) for the same radiology item are added in future.

search_exclude

A character vector of radiology names (Proc_Name in raw format) that are known NOT to be relevant.

value_as_character_fn

A function specifying how to generate the value_as_character column

x

Radiology data in tidy-column-name format (after applying radiology_rename)

errors

A function indicating what to do when an error is found

Value

A data frame with the following columns: person_id, symbol, value_as_number, value_as_character, censoring, administered_datetime,

unit, name, title, method, group, type

The result will be sorted by administered_datetime

Author

R.J.B. Goudie