Load data from Epic in delimited (including CSV) format

Convenience wrapper around readr::read_csv and readr::read_csv that makes it easier to set the col_types for standard data from Epic. Also makes handling untimezoned data more robust through use of functions from clock::parse_date_time.

Usage

read_tidybrookes_csv(
  file,
  col_types,
  n_max = Inf,
  na = c("", "NA"),
  quote = "\"",
  tz = "Europe/London",
  nonexistent = NULL,
  ambiguous = NULL
)

read_tidybrookes_delim(
  file,
  col_types,
  delim = "|",
  n_max = Inf,
  na = c("", "NA"),
  quote = "\""
)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector), as per readr::read_delim()

col_types

Either a string specifying the name of the table that the data comes from, for which a standard format can be used. Available values are adm, adt, demogs, fsheet, fsheet_io, tests, med_admin, med_prescr, diagnosis_pl, med_hist, radiology. This is translated into column specification using default_col_types.

Or a readr::cols() format list of column names and types, which can be used where a nonstandard data format are supplied.

n_max

Maximum number of lines to read.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

quote

Single character used to quote strings

tz

A timezone name, such as "Europe/London"

nonexistent

One of the following nonexistent time resolution strategies, allowed to be either length 1, or the same length as the input:

"roll-forward": The next valid instant in time.

"roll-backward": The previous valid instant in time.

"shift-forward": Shift the nonexistent time forward by the size of the daylight saving time gap.

"shift-backward: Shift the nonexistent time backward by the size of the daylight saving time gap.

"NA": Replace nonexistent times with NA.

"error": Error on nonexistent times.

Using either "roll-forward" or "roll-backward" is generally recommended over shifting, as these two strategies maintain the relative ordering between elements of the input.

If NULL, defaults to "error".

ambiguous

Passed to clock::date-time-parse, with the usual values being one of the following ambiguous time resolution strategies, allowed to be either length 1, or the same length as the input:

"earliest": Of the two possible times, choose the earliest one.

"latest": Of the two possible times, choose the latest one.

"NA": Replace ambiguous times with NA.

"error": Error on ambiguous times.

delim

Single character used to separate fields within a record.

Details

Named col_types are extracted using default_col_types().

At transitions to and from daylight savings times, there are nonexistent and ambiguous times. Using the parameters nonexistent and ambiguous the handling of these can be made explicit and consistent.

See also

Examples

adm_file_path <- tidybrookes_example("adm.csv")

# the format when col_types = "adm"
default_col_types("adm")
#> cols(
#>   STUDY_SUBJECT_DIGEST = col_character(),
#>   PAT_ENC_CSN = col_integer(),
#>   IN_DTTM = col_datetime(format = "%Y-%m-%d %H:%M:%S"),
#>   HOSP_DISCH_TIME = col_datetime(format = "%Y-%m-%d %H:%M:%S"),
#>   GENDER_DESC = col_character(),
#>   ETHNIC_GROUP_GROUPED = col_character(),
#>   DATE_OF_DEATH = col_datetime(format = "%Y-%m-%d %H:%M:%S"),
#>   AGE_AT_ADM = col_integer(),
#>   ADM_SERVICE = col_character(),
#>   ADT_DEPARTMENT_NAME = col_character(),
#>   DISCH_DEST = col_character(),
#>   DISCH_DECEASED = col_character(),
#>   READMIT_WITHIN_30 = col_character()
#> )

read_tidybrookes_csv(
  file = adm_file_path,
  col_types = "adm"
)
#> # A tibble: 5 × 13
#>   STUDY_SUBJECT_DIGEST PAT_ENC_CSN IN_DTTM             HOSP_DISCH_TIME     GENDER_DESC ETHNIC_GROUP_GROUPED DATE_OF_DEATH AGE_AT_ADM ADM_SERVICE        ADT_DEPARTMENT_NAME  DISCH_DEST DISCH_DECEASED
#>   <chr>                      <int> <dttm>              <dttm>              <chr>       <chr>                <dttm>             <int> <chr>              <chr>                <chr>      <chr>         
#> 1 AA                             1 2021-01-01 09:00:00 2021-01-03 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> 2 AA                             2 2021-01-05 11:00:00 2021-01-07 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT 0          1             
#> 3 BB                             3 2021-01-20 09:00:00 2021-02-01 18:00:00 Male        "White\n"            NA                    12 Nephrology         DIALYSIS CENTRE      1          0             
#> 4 CC                             4 2021-01-01 08:00:00 2021-02-01 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> 5 CC                             5 2021-02-02 08:00:00 2021-02-02 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> # ℹ 1 more variable: READMIT_WITHIN_30 <chr>

read_tidybrookes_csv(
  file = adm_file_path,
  col_types = "adm",
  tz = "Europe/London",
  nonexistent = "roll-forward",
  ambiguous = "latest"
)
#> # A tibble: 5 × 13
#>   STUDY_SUBJECT_DIGEST PAT_ENC_CSN IN_DTTM             HOSP_DISCH_TIME     GENDER_DESC ETHNIC_GROUP_GROUPED DATE_OF_DEATH AGE_AT_ADM ADM_SERVICE        ADT_DEPARTMENT_NAME  DISCH_DEST DISCH_DECEASED
#>   <chr>                      <int> <dttm>              <dttm>              <chr>       <chr>                <dttm>             <int> <chr>              <chr>                <chr>      <chr>         
#> 1 AA                             1 2021-01-01 09:00:00 2021-01-03 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> 2 AA                             2 2021-01-05 11:00:00 2021-01-07 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT 0          1             
#> 3 BB                             3 2021-01-20 09:00:00 2021-02-01 18:00:00 Male        "White\n"            NA                    12 Nephrology         DIALYSIS CENTRE      1          0             
#> 4 CC                             4 2021-01-01 08:00:00 2021-02-01 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> 5 CC                             5 2021-02-02 08:00:00 2021-02-02 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT 0          0             
#> # ℹ 1 more variable: READMIT_WITHIN_30 <chr>

# view default cols() format - this is what will be used if a character
# string name of a table is used, such as "fsheet"
default_col_types("fsheet")
#> cols(
#>   STUDY_SUBJECT_DIGEST = col_character(),
#>   fsd_id = col_integer(),
#>   line = col_integer(),
#>   `flo-meas_id` = col_character(),
#>   MEASURE_TIME = col_datetime(format = "%Y-%m-%d %H:%M:%S"),
#>   disp_name = col_character(),
#>   measured_value = col_character(),
#>   meas_comment = col_character(),
#>   template = col_character(),
#>   form = col_character()
#> )

# custom col_types can be specified
read_tidybrookes_csv(
  file = adm_file_path,
  col_types = cols(
    STUDY_SUBJECT_DIGEST = col_character(),
    fsd_id = col_character())
)
#> Warning: The following named parsers don't match the column names: fsd_id
#> # A tibble: 5 × 13
#>   STUDY_SUBJECT_DIGEST PAT_ENC_CSN IN_DTTM             HOSP_DISCH_TIME     GENDER_DESC ETHNIC_GROUP_GROUPED DATE_OF_DEATH AGE_AT_ADM ADM_SERVICE        ADT_DEPARTMENT_NAME  DISCH_DEST DISCH_DECEASED
#>   <chr>                      <dbl> <dttm>              <dttm>              <chr>       <chr>                <lgl>              <dbl> <chr>              <chr>                     <dbl>          <dbl>
#> 1 AA                             1 2021-01-01 09:00:00 2021-01-03 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> 2 AA                             2 2021-01-05 11:00:00 2021-01-07 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT          0              1
#> 3 BB                             3 2021-01-20 09:00:00 2021-02-01 18:00:00 Male        "White\n"            NA                    12 Nephrology         DIALYSIS CENTRE               1              0
#> 4 CC                             4 2021-01-01 08:00:00 2021-02-01 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> 5 CC                             5 2021-02-02 08:00:00 2021-02-02 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> # ℹ 1 more variable: READMIT_WITHIN_30 <dbl>

# Can use the clock package to determine handling of ambiguous and
# nonexistent dates (due to transitions to and from daylight saving time)
read_tidybrookes_csv(
  file = adm_file_path,
  col_types = cols(
    STUDY_SUBJECT_DIGEST = col_character(),
    fsd_id = col_character()),
  tz = "Europe/London",
  nonexistent = "roll-forward",
  ambiguous = "latest"
)
#> Warning: The following named parsers don't match the column names: fsd_id
#> # A tibble: 5 × 13
#>   STUDY_SUBJECT_DIGEST PAT_ENC_CSN IN_DTTM             HOSP_DISCH_TIME     GENDER_DESC ETHNIC_GROUP_GROUPED DATE_OF_DEATH AGE_AT_ADM ADM_SERVICE        ADT_DEPARTMENT_NAME  DISCH_DEST DISCH_DECEASED
#>   <chr>                      <dbl> <dttm>              <dttm>              <chr>       <chr>                <lgl>              <dbl> <chr>              <chr>                     <dbl>          <dbl>
#> 1 AA                             1 2021-01-01 09:00:00 2021-01-03 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> 2 AA                             2 2021-01-05 11:00:00 2021-01-07 18:00:00 Female      "White\n"            NA                    45 Emergency Medicine EMERGENCY DEPARTMENT          0              1
#> 3 BB                             3 2021-01-20 09:00:00 2021-02-01 18:00:00 Male        "White\n"            NA                    12 Nephrology         DIALYSIS CENTRE               1              0
#> 4 CC                             4 2021-01-01 08:00:00 2021-02-01 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> 5 CC                             5 2021-02-02 08:00:00 2021-02-02 18:00:00 Female      "White\n"            NA                    50 Emergency Medicine EMERGENCY DEPARTMENT          0              0
#> # ℹ 1 more variable: READMIT_WITHIN_30 <dbl>