Rename to tidy column names

Rename columns from the default names in the standard data format from Clinical Informatics into standardised, shorter names (e.g. all lowercase, underscore-separated, with _datetime suffix etc).

Usage

default_rename(x)

adm_rename(x, names = default_rename("adm"))

adt_rename(x, names = default_rename("adt"))

demogs_rename(x, names = default_rename("demogs"))

diagnosis_pl_rename(x, names = default_rename("diagnosis_pl"))

fsheet_io_rename(x, names = default_rename("fsheet_io"))

fsheet_rename(x, names = default_rename("fsheet"))

med_admin_rename(x, names = default_rename("med_admin"))

med_hist_rename(x, names = default_rename("med_hist"))

med_prescr_rename(x, names = default_rename("med_prescr"))

radiology_rename(x, names = default_rename("radiology"))

procedures_rename(x, names = default_rename("procedures"))

tests_rename(x, names = default_rename("tests"))

Arguments

x

A data frame of raw adm data

names

A vector of new_name = old_name pairs

Value

The supplied data frame, with column names in tidy-column-name format

Details

The default renaming map can be viewed using default_rename().

If variations from this format occur, custom renaming can be performed using the names argument.

See also

Author

R.J.B. Goudie

Examples

# The raw columns names are messy: long names, mixed capitalisation, etc!
fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)
#>  [1] "STUDY_SUBJECT_DIGEST" "fsd_id"               "line"                 "flo-meas_id"          "MEASURE_TIME"         "disp_name"            "measured_value"       "meas_comment"        
#>  [9] "template"             "form"                

# Default renaming can be viewed with
default_rename("fsheet")
#>              person_id                   name                  value                comment   measurement_datetime                data_id         measurement_id                line_id 
#> "STUDY_SUBJECT_DIGEST"            "disp_name"       "measured_value"         "meas_comment"         "MEASURE_TIME"               "fsd_id"          "flo-meas_id"                 "line" 
#>               template                   form 
#>             "template"                 "form" 

colnames(fsheet_rename(fsheet_imported_example))
#>  [1] "person_id"            "name"                 "value"                "comment"              "measurement_datetime" "data_id"              "measurement_id"       "line_id"             
#>  [9] "template"             "form"                

# A warning is issued if columns are missing
fsheet_imported_example %>%
  select(-MEASURE_TIME) %>%
  fsheet_rename
#> ! Some columns were missing: MEASURE_TIME
#> # A tibble: 32 × 9
#>    person_id name        value comment  data_id measurement_id line_id template  form       
#>    <chr>     <chr>       <chr> <chr>    <int64>        <int64> <int64> <chr>     <chr>      
#>  1 AA        Weight      4125  NA             5           1234      61 IP Vitals Vital signs
#>  2 AA        SpO2        96    NA      10909029             10      60 IP Vitals Vital signs
#>  3 AA        SpO2        90    NA      10909029             10      60 IP Vitals Vital signs
#>  4 AA        NEWS2 score 3     NA      10926333           1234     239 IP Vitals Vital signs
#>  5 AA        NEWS2 score 4     NA      10926333           1234     239 IP Vitals Vital signs
#>  6 AA        NEWS2 score 3     NA      10926333           1234     239 IP Vitals Vital signs
#>  7 AA        SpO2        96    NA      10909029             10      60 IP Vitals Vital signs
#>  8 AA        SpO2        90    NA      10909029             10      60 IP Vitals Vital signs
#>  9 AA        SpO2        92    NA      10909029             10      60 IP Vitals Vital signs
#> 10 AA        NEWS2 score 3     NA      10926333           1234     239 IP Vitals Vital signs
#> # ℹ 22 more rows