Renamed view of raw data to tidy column names

Create view of a table with standardised, shorter column names (e.g. all lowercase, underscore-separated, with _datetime suffix etc) rather than the default names in the standard data format from Clinical Informatics.

Usage

tbl_raw(src, table, names)

Arguments

src

A DBIConnection object produced by duckdb_tidybrookes_connect or DBI::dbConnect()

table

A string specifying the name of the table. Typical values would be adm, adt, demogs, fsheet, fsheet_io, tests, med_admin, med_prescr, diagnosis_pl, med_hist, radiology

names

A vector of new_name = old_name pairs

Details

The default renaming map can be viewed using default_rename().

If variations from this format occur, custom renaming can be performed using the names argument.

Author

R.J.B. Goudie