Add annotations to tests data

Annotate tests data to note rows that should be excluded etc.

Usage

tests_annotate(x, tests_def, annotation_db = NULL)

Arguments

x

Tests data in renamed format (after applying tests_rename)

tests_def

A tests definition, or list of tests definitions

annotation_db

Either NULL or a tbl reference to the annotations database table that the annotations should be stored in. If NULL the annotations are added to x as additional columns.

Details

Annotations can either be added as additional columns (when annotation_db = NULL), or stored in a separate database table (by setting annotation_db to a database tbl).

Note that when adding the annotations as additiona columns, the result will not be sorted, since it is left in the original order of x

Author

R.J.B. Goudie