Add a new definition of a test item

Add a new definition of a test item

Extract tests data into tidy format

Usage

tests_add(
  test_def,
  symbol,
  title,
  names_cuh = NA,
  names_external = NA,
  search_pattern,
  search_exclude = NA,
  search_exclude_group = NA,
  type = "numeric",
  silently_exclude_na_when = FALSE,
  silently_exclude_when = FALSE,
  censoring_fn = case_when(TRUE ~ NA_character_),
  value_as_number_fn = if_else(type == "numeric", case_when(TRUE ~ value_as_number),
    case_when(TRUE ~ NA_real_)),
  value_as_character_fn = if_else(type == "character", case_when(TRUE ~
    value_as_character), case_when(TRUE ~ NA_character_)),
  value_as_logical_fn = if_else(type == "logical", case_when(TRUE ~ value_as_logical),
    case_when(TRUE ~ as.logical(NA))),
  unit_rescale_fn = case_when(TRUE ~ value_as_number),
  unit_relabel_fn = case_when(TRUE ~ unit),
  expect_before = TRUE,
  expect_after = TRUE,
  range_mainly_low = NA,
  range_mainly_high = NA,
  range_discard_below = NA,
  range_discard_above = NA
)

tests_extract(
  x,
  tests_def,
  errors = stop,
  rds_only = FALSE,
  rds_filepath_fn = NULL
)

Arguments

test_def

A list of existing test definitions to add this to

symbol

A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name

title

The full name for the test item, suitable for using in figure titles etc

names_cuh

The names of the test (TestName in raw format) for tests to include from CUH

names_external

The names of the test (TestName in raw format) for tests to include from external sources. NOTE CURRENTLY WE DON'T INCLUDE EXTERNAL TESTS AT ALL.

search_pattern

A charater vector of potential synonymns for this test, in case new names (TestName in raw format) for the same test are added in future.

search_exclude

A character vector of test names (TestName in raw format) that are known NOT to be relevant to this test.

search_exclude_group

A character vector of test group names (TestGroupName in raw format) that are known NOT to be relevant to this test.

type

Whether flowsheet item is a number or a string. Either "numeric" or "character"

silently_exclude_na_when

An expression that returns a logical value, specifying the circumstances when NA values should be automatically excluded.

silently_exclude_when

An expression that returns a logical value, specifying the circumstances when values should be automatically excluded

censoring_fn

A function specifying the value of the censoring column

value_as_number_fn

A function specifying how to generate the value_as_number column

value_as_character_fn

A function specifying how to generate the value_as_character column

value_as_logical_fn

A function specifying how to generate the value_as_logical column

unit_rescale_fn

A function specifying how to rescale the result values. By default, does not rescale.

unit_relabel_fn

A function specifying how to relabel the units of the the result values. By default, does not rescale.

expect_before

An expression that returns a logical value, specifying a condition that should be TRUE in the raw data

expect_after

An expression that returns a logical value, specifying a condition that should be TRUE in the processed (output) data

range_mainly_low, range_mainly_high

An indicative lower and upper range for most values. Values outside this range are NOT excluded: this is purely for setting default scales of plots etc

range_discard_below, range_discard_above

This pair set the lower and upper range for the flowsheet item. Values outside this range are EXCLUDED.

x

Test data in tidy-column-name format (after applying tests_rename)

tests_def

A tests definition

errors

A function indicating what to do when an error is found

Value

A data frame with the following columns: person_id, symbol, value_as_number, value_as_character, value_as_logical, censoring, ordered_datetime, collected_datetime, received_datetime, result_datetime, ordering_department_name, range_low, range_high, unit, name, title, method, group, type

The result will be sorted by collected_datetime

Author

R.J.B. Goudie

Examples

tests_raw_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("tests.csv"),
    col_types = "tests") %>%
    tests_rename

tests_def <- list() %>%
  tests_add(
    symbol = "magnesium",
    title = "Magnesium",
    names_cuh = c("MAGNESIUM"),
    names_external = "EXT MAGNESIUM",
    search_pattern = c("magnesium", "mg"),
    search_exclude = c("ETOH MG/100ML",
                       "RETIRED-ETOH MG/100ML",
                       "RIST 0.2MG/ML",
                       "RIST 0.77MG/ML",
                       "RISTOCETIN 0.5MG/ML",
                       "RISTOCETIN 1.5MG/ML"),
    silently_exclude_na_when = FALSE,
    censoring_fn = case_when(value_original == "<0.29" ~ "left", # very low
                             TRUE ~ NA_character_),
    value_as_number_fn = case_when(value_original == "<0.29" ~ 0.29,
                                   TRUE ~ value_as_number),
    expect_before = (unit == "mmol/L"),
    range_mainly_low = 0.4,
    range_mainly_high = 1.2,
    range_discard_above = 1.8)

tests_raw_example %>% filter(name == "MAGNESIUM")
#> # A tibble: 9 × 14
#>   person_id name      value ordered_datetime    collected_datetime  received_datetime   result_datetime     ordering_department  range_low range_high unit   method group    order_id
#>   <chr>     <chr>     <chr> <dttm>              <dttm>              <dttm>              <dttm>              <chr>                <chr>     <chr>      <chr>  <chr>  <chr>    <chr>   
#> 1 AA        MAGNESIUM 0.9   2021-01-02 09:00:00 2021-01-02 09:40:00 2021-01-02 09:50:00 2021-01-02 18:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 2 AA        MAGNESIUM 3.0   2021-01-03 09:00:00 2021-01-03 09:40:00 2021-01-03 09:50:00 2021-01-03 18:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 3 BB        MAGNESIUM 0.8   2021-01-19 09:00:00 2021-01-19 08:40:00 2021-01-19 09:40:00 2021-01-19 12:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 4 BB        MAGNESIUM 0.9   2021-01-20 09:00:00 2021-01-20 08:40:00 2021-01-20 09:40:00 2021-01-20 12:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 5 BB        MAGNESIUM 1.0   2021-01-21 09:00:00 2021-01-21 08:40:00 2021-01-21 09:40:00 2021-01-21 12:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 6 BB        MAGNESIUM 1.2   2021-01-22 09:00:00 2021-01-22 08:40:00 2021-01-22 09:40:00 2021-01-22 12:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 7 BB        MAGNESIUM 0.7   2021-01-23 09:00:00 2021-01-23 08:40:00 2021-01-23 09:40:00 2021-01-23 12:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 8 CC        MAGNESIUM 0.61  2021-02-01 14:00:00 2021-02-01 15:40:00 2021-02-01 16:50:00 2021-02-01 17:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    
#> 9 CC        MAGNESIUM <0.29 2021-02-02 14:00:00 2021-02-02 15:40:00 2021-02-02 16:50:00 2021-02-02 17:00:00 EMERGENCY DEPARTMENT 0.6       1          mmol/L ABC    MINERALS 1242    

# Note that
# - the high value of 3.0 is deleted because it is above
#   range_discard_above
# - note that the censored value "<0.29" is converted to 0.29, and "left"
#   added to the censoring column
tests_extract(tests_raw_example, tests_def)
#> ℹ Extracting Magnesium (magnesium) from 9 raw rows
#> ℹ 1 row removed since >1.8
#> ℹ 8 rows extracted
#> # A tibble: 8 × 20
#>   person_id symbol    value_as_character value_as_number censoring value_as_logical ordered_datetime    collected_datetime  received_datetime   result_datetime     ordering_department  range_low
#>   <chr>     <chr>     <chr>                        <dbl> <chr>     <lgl>            <dttm>              <dttm>              <dttm>              <dttm>              <chr>                <chr>    
#> 1 AA        magnesium NA                            0.9  NA        NA               2021-01-02 09:00:00 2021-01-02 09:40:00 2021-01-02 09:50:00 2021-01-02 18:00:00 EMERGENCY DEPARTMENT 0.6      
#> 2 BB        magnesium NA                            0.8  NA        NA               2021-01-19 09:00:00 2021-01-19 08:40:00 2021-01-19 09:40:00 2021-01-19 12:00:00 EMERGENCY DEPARTMENT 0.6      
#> 3 BB        magnesium NA                            0.9  NA        NA               2021-01-20 09:00:00 2021-01-20 08:40:00 2021-01-20 09:40:00 2021-01-20 12:00:00 EMERGENCY DEPARTMENT 0.6      
#> 4 BB        magnesium NA                            1    NA        NA               2021-01-21 09:00:00 2021-01-21 08:40:00 2021-01-21 09:40:00 2021-01-21 12:00:00 EMERGENCY DEPARTMENT 0.6      
#> 5 BB        magnesium NA                            1.2  NA        NA               2021-01-22 09:00:00 2021-01-22 08:40:00 2021-01-22 09:40:00 2021-01-22 12:00:00 EMERGENCY DEPARTMENT 0.6      
#> 6 BB        magnesium NA                            0.7  NA        NA               2021-01-23 09:00:00 2021-01-23 08:40:00 2021-01-23 09:40:00 2021-01-23 12:00:00 EMERGENCY DEPARTMENT 0.6      
#> 7 CC        magnesium NA                            0.61 NA        NA               2021-02-01 14:00:00 2021-02-01 15:40:00 2021-02-01 16:50:00 2021-02-01 17:00:00 EMERGENCY DEPARTMENT 0.6      
#> 8 CC        magnesium NA                            0.29 left      NA               2021-02-02 14:00:00 2021-02-02 15:40:00 2021-02-02 16:50:00 2021-02-02 17:00:00 EMERGENCY DEPARTMENT 0.6      
#> # ℹ 8 more variables: range_high <chr>, unit <chr>, name <chr>, title <chr>, method <chr>, group <chr>, order_id <chr>, type <chr>