Widen test data

Reshapes the data so that all tests from the same order_id for the same person appear in columns

Usage

tests_pivot_wider_order(x, values_fn = NULL, include_censoring = FALSE)

Arguments

x

A tidy test data frame

values_fn

A function to summarise multiple results for a particular test from a particular order.

include_censoring

Logical, should censoring column be included?

Value

A data frame with the tests in columns

Author

R.J.B. Goudie