Rename from tidy to raw column names

The standard data format from Clinical Informatics is handled by default. If variations from this format occur, custom renaming can be performed using the names argument

Usage

adm_unrename(x, names = flip_names_and_values(default_rename("adm")))

adt_unrename(x, names = flip_names_and_values(default_rename("adt")))

demogs_unrename(x, names = flip_names_and_values(default_rename("demogs")))

diagnosis_pl_unrename(
  x,
  names = flip_names_and_values(default_rename("diagnosis_pl"))
)

fsheet_io_unrename(
  x,
  names = flip_names_and_values(default_rename("fsheet_io"))
)

fsheet_unrename(x, names = flip_names_and_values(default_rename("fsheet")))

med_admin_unrename(
  x,
  names = flip_names_and_values(default_rename("med_admin"))
)

med_prescr_unrename(
  x,
  names = flip_names_and_values(default_rename("med_prescr"))
)

radiology_unrename(
  x,
  names = flip_names_and_values(default_rename("radiology"))
)

procedures_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", PROCEDUREDATE = "procedure_datetime",
    PROCEDUREDESC = "description", OPCSCODE = "opcs")
)

tests_unrename(
  x,
  names = c(STUDY_SUBJECT_DIGEST = "person_id", TestName = "name", ResultValue = "value",
    ORDERED_DATETIME = "ordered_datetime", COLLECTED_DATETIME = "collected_datetime",
    RECEIVED_DATETIME = "received_datetime", ResultDate = "result_datetime",
    ORDERING_DEPARTMENT_NAME = "ordering_department", ReferenceLow = "range_low",
    ReferenceHigh = "range_high", ResultUnit = "unit", Method = "method", TestGroupName =
    "group", OrderProcId = "order_id")
)

Arguments

x

A data frame of raw adm data with tidy names

names

A vector of new_name = old_name pairs

Value

The supplied data frame, with column names in tidy-column-name format

See also

Author

R.J.B. Goudie

Examples

fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)
#>  [1] "STUDY_SUBJECT_DIGEST" "fsd_id"               "line"                 "flo-meas_id"          "MEASURE_TIME"         "disp_name"            "measured_value"       "meas_comment"        
#>  [9] "template"             "form"                
colnames(fsheet_rename(fsheet_imported_example))
#>  [1] "person_id"            "name"                 "value"                "comment"              "measurement_datetime" "data_id"              "measurement_id"       "line_id"             
#>  [9] "template"             "form"                
colnames(fsheet_unrename(fsheet_rename(fsheet_imported_example)))
#>  [1] "STUDY_SUBJECT_DIGEST" "disp_name"            "measured_value"       "meas_comment"         "MEASURE_TIME"         "fsd_id"               "flo-meas_id"          "line"                
#>  [9] "template"             "form"