Join adm data with other (standardised) data

Joins an adm data frame with another data frame (with a person_id for linking and with the relevant datetime called datetime), and filter to only those rows of y with datetime within the specified time period.

Usage

all_during(x, y, datetime, during, names_from = "symbol", join = "left")

Arguments

x

An adm data frame

y

A (tidy) data frame, with a datetime column

datetime

The column of y to use as the main datetime for matching

during

The time period to extract data for, one of:

"during_visit"

"during_visit_initial_24h"

"during_visit_initial_72h

"during_icu"

"before_visit_initial_24h"

"14_days_before_visit_until_visit_end"

"year_before_visit_until_visit_end"

"before_visit_end"

"30_days_before_visit_start"

"year_before_initial_24h"

"during_value"

names_from

A character vector containing the names of columns that should be used to filter by (in addition to those implied by during).

join

Either "left", in which a left join is performed (so all x rows will be returned, even if they have no correspoding y) or "inner", in which an inner join is performed (so only x rows with at least one y row will be returned)

Value

A data frame, with a row for each y measurement during the relevant time period for each patient. Note that rows of the adm data will be repeated multiple times (one for each y measurement data point).

See also

Author

R.J.B. Goudie

Examples

adm_data_example
#> # A tibble: 5 × 13
#>   person_id visit_id visit_start_datetime visit_end_datetime  gender ethnicity death_date age_at_visit_start adm_service        ward    discharge_destination discharged_deceased readmitted_within_30…¹
#>   <chr>        <int> <dttm>               <dttm>              <chr>  <chr>     <dttm>                  <int> <chr>              <chr>   <chr>                 <chr>               <chr>                 
#> 1 AA               1 2021-01-01 09:00:00  2021-01-03 18:00:00 Female "White\r… NA                         45 Emergency Medicine EMERGE… 0                     0                   1                     
#> 2 AA               2 2021-01-05 11:00:00  2021-01-07 18:00:00 Female "White\r… NA                         45 Emergency Medicine EMERGE… 0                     1                   0                     
#> 3 BB               3 2021-01-20 09:00:00  2021-02-01 18:00:00 Male   "White\r… NA                         12 Nephrology         DIALYS… 1                     0                   0                     
#> 4 CC               4 2021-01-01 08:00:00  2021-02-01 18:00:00 Female "White\r… NA                         50 Emergency Medicine EMERGE… 0                     0                   1                     
#> 5 CC               5 2021-02-02 08:00:00  2021-02-02 18:00:00 Female "White\r… NA                         50 Emergency Medicine EMERGE… 0                     0                   0                     
#> # ℹ abbreviated name: ¹​readmitted_within_30_days
fsheet_news2_example
#> # A tibble: 12 × 17
#>    person_id symbol measurement_datetime value_as_number censoring value_as_logical value_as_character name        unit  type    template  form         data_id line_id measurement_id title comment
#>    <chr>     <chr>  <dttm>                         <dbl> <chr>     <lgl>            <chr>              <chr>       <chr> <chr>   <chr>     <chr>          <int>   <int> <chr>          <chr> <chr>  
#>  1 CC        news2  2021-01-01 08:59:00                5 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  2 AA        news2  2021-01-01 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  3 AA        news2  2021-01-01 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  4 AA        news2  2021-01-01 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  5 CC        news2  2021-01-03 17:00:00                7 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  6 CC        news2  2021-01-04 09:00:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  7 AA        news2  2021-01-05 11:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  8 AA        news2  2021-01-05 16:30:00                6 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#>  9 AA        news2  2021-01-06 10:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 10 BB        news2  2021-01-21 09:30:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 11 BB        news2  2021-01-22 12:30:00                4 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     
#> 12 BB        news2  2021-01-23 09:00:00                3 NA        NA               NA                 NEWS2 score NA    numeric IP Vitals Vital signs 10926333     239 1234           NEWS2 NA     

all_during(adm_data_example,
           fsheet_news2_example,
           datetime = measurement_datetime,
           during = "during_visit") %>%
  select(person_id, visit_id, symbol, datetime, value_as_number)
#> # A tibble: 13 × 5
#>    person_id visit_id symbol datetime            value_as_number
#>    <chr>        <int> <chr>  <dttm>                        <dbl>
#>  1 AA               1 news2  2021-01-01 09:00:00               3
#>  2 AA               1 news2  2021-01-01 09:30:00               3
#>  3 AA               1 news2  2021-01-01 12:30:00               4
#>  4 AA               2 news2  2021-01-05 11:30:00               3
#>  5 AA               2 news2  2021-01-05 16:30:00               6
#>  6 AA               2 news2  2021-01-06 10:30:00               3
#>  7 BB               3 news2  2021-01-21 09:30:00               3
#>  8 BB               3 news2  2021-01-22 12:30:00               4
#>  9 BB               3 news2  2021-01-23 09:00:00               3
#> 10 CC               4 news2  2021-01-01 08:59:00               5
#> 11 CC               4 news2  2021-01-03 17:00:00               7
#> 12 CC               4 news2  2021-01-04 09:00:00               6
#> 13 CC               5 NA     NA                               NA