Add a new definition of a medication administration drug

Add a new definition of a medication administration drug

Extract med_admin data into tidy format

Usage

med_admin_add(
  med_admin_def,
  note,
  symbol,
  title,
  names = NA,
  search_pattern,
  search_exclude = NA,
  silently_exclude_na_routes = FALSE,
  route_class = "intravenous",
  route_include = default_route_include(route_class),
  route_exclude = default_route_exclude(route_class),
  action_class = c("bolus", "infusion", "pca"),
  action_include = default_action_include(action_class),
  action_exclude = default_action_exclude(action_class),
  concentration_fn = case_when(TRUE ~ NA_real_),
  dose_range_discard_above_fn = case_when(TRUE ~ Inf),
  dose_range_mainly_below_fn = NA,
  expect_after = TRUE
)

med_admin_extract(x, med_admin_def, med_admin_units, errors = stop)

Arguments

med_admin_def

A med_admin definition

note

A character vector of notes that should be printed

symbol

A short name for the flowsheet item. This should be a character vector that is suitable to use as an R object name

title

The full name for the test item, suitable for using in figure titles etc

names

The names of the Medications (DrugName in raw format)

search_pattern

A charater vector of potential synonymns for this drug, in case new names (DrugName in raw format) for the same drug are added in future.

search_exclude

A character vector of drug names (DrugName in raw format) that are known NOT to be relevant to this drug

silently_exclude_na_routes

An logical specifying whether medication administration records with NA routes should be automatically excluded.

route_class

One of enteral, intravenous, other_systemic, respiratory or topical. These classes will use the classification in med_admin_routes.csv to specify a list of routes to include

route_include

A vector of route names (RouteOfMedicationAbbreviated in raw format) to include

route_exclude

A vector of route names (RouteOfMedicationAbbreviated in raw format) to exclude

action_class

One of infusion, bolus, pca, feed. These classes will use the classification in med_admin_action.csv to specify a list of actions to include

action_include

A vector of action names (MARAction in raw format) to include

action_exclude

A vector of action names (MARAction in raw format) to exclude

concentration_fn

A function that codes the concentration of each row of the medication administration records (for that drug). It should return a real. This will typically be done on the basis of the name, info_concentration, info_conversion or info_rate, but can use any variables in the dataframe.

dose_range_mainly_below_fn

NOT CURRRENTLY IMPLEMENTED. An indicative lower and upper range for most values. Values outside this range are NOT excluded: this is purely for setting default scales of plots etc

expect_after

An expression that returns a logical value, specifying a condition that should be TRUE in the processed (output) data

x

Medication administration data in tidy-column-name format (after applying med_admin_rename)

errors

A function indicating what to do when an error is found

dose_range_discard_above_fne

NOT CURRRENTLY IMPLEMENTED. This seta the upper range for the dose item. Values outside this range are EXCLUDED.

Value

A data frame with the following columns: person_id, symbol, value_as_number, value_as_character, censoring, administered_datetime,

unit, name, title, method, group, type

The result will be sorted by administered_datetime

Author

R.J.B. Goudie