Rename to tidy column names

Rename columns from the default names in the standard data format from Clinical Informatics into standardised, shorter names (e.g. all lowercase, underscore-separated, with _datetime suffix etc).

Usage

default_rename(x)

adm_rename(x, names = default_rename("adm"))

adt_rename(x, names = c())

demogs_rename(x, names = c())

diagnosis_pl_rename(x, names = c())

fsheet_io_rename(x, names = c())

fsheet_rename(x, names = c())

med_admin_rename(x, names = c())

med_hist_rename(x, names = c())

med_prescr_rename(x, names = c())

radiology_rename(x, names = c())

procedures_rename(x, names = c())

tests_rename(x, names = c())

Arguments

x

A data frame of raw adm data

names

A vector of new_name = old_name pairs that will override the default renaming (the default renaming is merged with these names, with these names taking priority)

Value

The supplied data frame, with column names in tidy-column-name format

Details

The default renaming map can be viewed using default_rename().

If variations from this format occur, custom renaming can be performed using the names argument.

See also

Author

R.J.B. Goudie

Examples

# The raw columns names are messy: long names, mixed capitalisation, etc!
fsheet_imported_example <-
  read_tidybrookes_csv(
    file = tidybrookes_example("fsheet.csv"),
    col_types = "fsheet"
  )
colnames(fsheet_imported_example)
#>  [1] "STUDY_SUBJECT_DIGEST" "fsd_id"               "line"                 "flo-meas_id"          "MEASURE_TIME"         "disp_name"            "measured_value"       "meas_comment"        
#>  [9] "template"             "form"                

# Default renaming can be viewed with
default_rename("fsheet")
#>              person_id                   name                   name                  value                  value                comment   measurement_datetime                data_id 
#> "STUDY_SUBJECT_DIGEST"            "disp_name"            "DISP_NAME"       "measured_value"       "MEASURED_VALUE"         "meas_comment"         "MEASURE_TIME"               "fsd_id" 
#>         measurement_id         measurement_id                line_id               template               template                   form                   form 
#>          "flo-meas_id"          "FLO_MEAS_ID"                 "line"             "template"             "TEMPLATE"                 "form"                 "FORM" 

colnames(fsheet_rename(fsheet_imported_example))
#>  [1] "person_id"            "name"                 "value"                "comment"              "measurement_datetime" "data_id"              "measurement_id"       "line_id"             
#>  [9] "template"             "form"                

# Provided names override the defaults
fsheet_imported_example %>%
  fsheet_rename(names = c(custom_person_id = "STUDY_SUBJECT_DIGEST"))
#> # A tibble: 32 × 10
#>    custom_person_id name        value comment measurement_datetime  data_id measurement_id line_id template  form       
#>    <chr>            <chr>       <chr> <chr>   <dttm>                <int64>        <int64> <int64> <chr>     <chr>      
#>  1 AA               Weight      4125  NA      2021-01-03 17:00:00         5           1234      61 IP Vitals Vital signs
#>  2 AA               SpO2        96    NA      2021-01-01 10:59:00  10909029             10      60 IP Vitals Vital signs
#>  3 AA               SpO2        90    NA      2021-01-02 08:59:00  10909029             10      60 IP Vitals Vital signs
#>  4 AA               NEWS2 score 3     NA      2021-01-01 09:30:00  10926333           1234     239 IP Vitals Vital signs
#>  5 AA               NEWS2 score 4     NA      2021-01-01 12:30:00  10926333           1234     239 IP Vitals Vital signs
#>  6 AA               NEWS2 score 3     NA      2021-01-01 09:00:00  10926333           1234     239 IP Vitals Vital signs
#>  7 AA               SpO2        96    NA      2021-01-05 11:30:00  10909029             10      60 IP Vitals Vital signs
#>  8 AA               SpO2        90    NA      2021-01-05 16:30:00  10909029             10      60 IP Vitals Vital signs
#>  9 AA               SpO2        92    NA      2021-01-06 10:30:00  10909029             10      60 IP Vitals Vital signs
#> 10 AA               NEWS2 score 3     NA      2021-01-05 11:30:00  10926333           1234     239 IP Vitals Vital signs
#> # ℹ 22 more rows